[Fuchsia] Write net-test-server-config to /tmp
Previuously net-test-server-config was written to /data/shared,
which is then mapped to /test-shared in the test launcher. The problem
was that this approach breaks when test launcher is not used (i.e. with
--single-process-tests flag). Updated runner scripts to copy the file to
/tmp. This approach is more consistent with other platforms and doesn't
require any special logic in the test launcher.

Change-Id: I5377eda57a0524b02e7ce419a5930ff16087a2e6
Reviewed-on: https://chromium-review.googlesource.com/c/1374456
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: Kevin Marshall <kmarshall@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Sergey Ulanov <sergeyu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#616052}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9b25ed223d2502b9f6a99dea903a5b95376a2d51