Bound the number of outstanding writes to SSLKEYLOGFILE.
This is a reland of
https://chromium-review.googlesource.com/c/chromium/src/+/1377187, with
a fix to not call base::Bind in the constructor.

Some AVs appear to be using this file, but pointing it to a pipe where
they cannot keep up with writes. As we should not build up an unbounded
buffer when I/O is slow, just drop the entries past some very generous
limit.

Tested manually by bringing the limit down to 5 and pointing
SSLKEYLOGFILE at a named pipe attached to a reader I had suspended (at
which point it still took a lot of work to reach that limit).

Bug: 566951, 914880
Change-Id: I0ae1b6de548a9a345d9dc83b509062b2cb0aaa62
Reviewed-on: https://chromium-review.googlesource.com/c/1378639
Reviewed-by: Adam Langley <agl@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617210}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d054e43739f1e8e462ab03760d7ea168fa3a7dab