Ignore CreationDate when deleting a cookie
It is possible that the CookieMonster modifies the creation timestamp of
a Cookie during creation or updating and the CookiesTree UI doesn't
get notified about this. If a user later tries to delete a cookie
with a timestamp that is not consistent with the timestamp in the cookie
store, this deletion would fail.
To fix the issue, the timestamp is ignored. As we still check cookies
for being equivalent and having the same value, the property that
updated cookies shouldn't get deleted is preserved.

Bug: 918413
Change-Id: I392e19219c612021003ef7abe99bc52cdbb23173
Reviewed-on: https://chromium-review.googlesource.com/c/1396119
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#620751}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f18b096e2de34e724331bfa7b8d3b25eb2e1fc8e