[Fuchsia] Fix SocketPosix::IsConnected() pre-Connect() behaviour.
SocketPosix::IsConnected() has a Fuchsia-specific implementation, to
work-around the lack of a working recvmsg(MSG_PEEK). This was
incorrectly returning |true| when a SocketPosix was not-yet-connected.

Fix the Fuchsia implementation to report |false| in this case, and
re-enable a TCPClientSocket test that depends on it.

Add a TCPClientSocket test for the behaviour of both the IsConnected()
and IsConnectedAndIdle() APIs.

Bug: 887587
Change-Id: I5b7ef9fa7f98364ee5612cf8d472a13ebbeab164
Reviewed-on: https://chromium-review.googlesource.com/c/1407505
Commit-Queue: Wez <wez@chromium.org>
Reviewed-by: Sergey Ulanov <sergeyu@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#622231}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 372b03421362c7e14298e471451fd76223d0780d