OpenOrCreateEntry returns if entry was opened
disk_cache::OpenOrCreateEntry() previously did not indicated if an
entry was opened or created when called. This CL creates a new struct
containing both the Entry* and a bool indicating if the action
performed was opening or not.

Change-Id: Iabbe11a6feb4500aa9ff297e62be024bb1b441c6
Reviewed-on: https://chromium-review.googlesource.com/c/1409730
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#623015}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6a1d77d3d73cf4538dce28d8d168ac411c54a08e