Try to fix TransportClientSocketTest.FullDuplex_WriteFirst.
At the end of the test it closes the other end of socket that it has a
pending write to, then reads from the socket, expecting success or a
socket closure, but apparently can get a connection reset on Win64.
My guess is what's happening is the socket is closed before the pending
write completes, and writing to the closed socket then results in a
connection reset.

Regardless of the cause, though, there's no actual need to close the
socket - data was successfully written to the socket by this point, and
the socket was never read from, so the read should succeed without
closing the socket, contrary to what a comment claims. It looks to me
like the comment has been wrong since the CL landed.

Bug: 552053
Change-Id: Ieb64f71ad7c54ec6b5be455ffeaf98ed3a7c2b59
Reviewed-on: https://chromium-review.googlesource.com/c/1450431
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#628442}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 08b33971acef4b429229b2aaea362449beeb64e5