Socket Pools Refactor 16 (3): Remove two of the SSL socket pools.
The CL removes the main SSL socket pool, merging it with the
TransportClientSocketPool below it. It does the same for the SSL socket
pool used for HTTPS proxies. The order there was:
Host SSL pool -> http proxy pool -> proxy SSL pool -> transport pool.
This CL merges the last two in that list.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: I9e94d7a257e20b74eb1399b01a4112038cb5beda
Reviewed-on: https://chromium-review.googlesource.com/c/1448993
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#628989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d5e2c9e9f86609c88fe102b92df5a395ed6afa1