Socket Pools Refactor 17: Move SSL tagging tests to transport test file.
The SSLClientSocketPool is going away, and the TransportClientSocketPool
is going to be used instead. This CL moves socket tagging tests (Which
were alread switching to using TransportClientSocketPool in an earlier
CL) to the same file as the other TransportClientSocketPool test.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: I6996e16b98ad78a92149beaeb5d62543fdecbff6
Reviewed-on: https://chromium-review.googlesource.com/c/1443733
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Paul Jensen <pauljensen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#629010}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f045145ab566e985536adc01ac4ef64734287c91