Socket Pools Refactor 21: Make HttpStreamParser use StreamSockets.
Previously it used ClientSocketHandles.

This change will make it easier to have it sit on top of a socket that
is not in a socket pool.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: I77507a408ed5bcf1546f903e7fb7c82f7bb8a7c5
Reviewed-on: https://chromium-review.googlesource.com/c/1450864
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#629279}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f171ff8bb59bbdfcf9f145c4d360c857b91c0bdb