Fix MockSimplePersistentCookieStore to use the proper key.
As in one that follows from cookie semantics. We previously tried to use
Creation timestamp as the key, with weird hacks to make it unique...
but that's mostly gone now. This removes one of the simpler remnants.

Change-Id: I10670abff8976ec8ee941fe42e4247cf1218d751
Reviewed-on: https://chromium-review.googlesource.com/c/1455297
Reviewed-by: Misha Efimov <mef@chromium.org>
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#629341}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4883bf37d776f559f45c8d8be9c2c26c2aca71ee