Socket Pools Refactor 19: Move HttpProxyConnectJob to its own file.
And rewrite most HttpProxyClientSocketPool tests in terms of it. The
only remaining tests are for HTTP auth, client certs, socket tagging,
or for cases that return a socket on failure, all of which include
some aspect of testing the socket pool as well, and will be delth with
in future CLs.

This CL also changes type of the |respect_limits| argument to
HttpProxyConnectJob to a bool, from a ClientSocketPool::RespectLimits.
HttpProxyConnectJob is otherwise unchanged.

The HttpProxyClientSocketPool class is going to be merged into the
TransportClientSocketPool class, but we'll still need these tests.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: Ic0570a7e48ad012da793723bbecbdf04d3659586
Reviewed-on: https://chromium-review.googlesource.com/c/1446563
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#630576}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 47a8ec134c4b6f01bfc3b99939ae3d5de70c4ff3