Socket Pools Refactor 20: Make HttpProxyConnectJob use common parameters
This CL makes HttpProxyConnectJob  use CommonConnectJobParams, adding
ProxyDelegate to the CommonConnectJobParams object. This is the last
ConnectJob type to use the struct, so this CL also removes the legacy
ConnectJob constructor.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: I4d3e4b99469ddb9a275e0668b4c827f405a73a3d
Reviewed-on: https://chromium-review.googlesource.com/c/1446790
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#630600}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1751ba7daafabba8ed50ff90945137ca3fd24b80