Remove custom HostResolver mock in http tests
Added functionality to MockCachingHostResolver to handle automatic
invalidation of cache entries after X resolves from cache.

As this was the last override of MockHostResolverBase, convert the
constructor to private with friends to force everything to deal with
MockHostResolver or MockCachingHostResolver as that seems to be the
intention of MockHostResolverBase.

Bug: 922699
Change-Id: I245e9ee32683f67cfcc7b1ad9a406cfbd223f56e
Reviewed-on: https://chromium-review.googlesource.com/c/1478354
Auto-Submit: Eric Orth <ericorth@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#633870}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1da75defee92ee59ffb8c9749c5d9de8c872d120