Make StdStringCanonOutput non copyable and don't copy it
In some build experiments, the compiler emitted an object assignment
of StdStringCanonOutput which did not link because the operator=
was not exported from the url.dll in component builds.

This makes the class  explicitly not copyable and removes
the copy to avoid that kind of (hard to debug) problems.

Change-Id: I92988b5799e4229d1fb0379f9897727c1bac20de
Reviewed-on: https://chromium-review.googlesource.com/c/1477602
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Commit-Queue: Ryan Hamilton <rch@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#634383}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ae55224516c70ffa0ee1a030f591eea44f22c5ab