Socket Pools Refactor 27: Rework HTTP proxy auth (part 1).
This CL adds a new callback passed in to ClientSocketPools when making
a request, to handle HTTP auth. The callback is never invoked, nor is
a non-NULL callback ever passed in, in this CL. This CL just updates
all consumers to pass pass in an empty callback, and
ClientSocketPoolBase to track the callback for the request.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: Icd4b0024d3fb83d3f1e452d817a0361d83c922dd
Reviewed-on: https://chromium-review.googlesource.com/c/1466869
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#635286}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 28ac03e67a807d5513d281ea98d39fd1166db1a7