Split Cache: Privacy preserving interpretation of empty top frame origin
If a request comes without a top frame origin, then skip the caching of that
resource or using the cache for accessing that resource. This is to make any
privacy attacks not dependent on the context that initiated the resource
request.

Test: net_unittests --gtest_filter=*SplitCache*
Bug: 936034
Change-Id: I92c54d806b6eadc833af4f8a0235259143a79d28
Reviewed-on: https://chromium-review.googlesource.com/c/1489531
Commit-Queue: Shivani Sharma <shivanisha@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#636009}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c7bbfba4fa2fb01aa94cc2abc962fcba4df6892e