Socket Pools Refactor 30: Finish flattening the socket pools.
This CL merges the SSL socket pool that sits on top of the
HTTP proxy socket pool into that socket pool. SSLConnectJobs
that are layered on top of HttpProxyConnectJobs now directly
create those nested jobs, instead of going through another
socket pool.

While this is the final CL needed to fully flatten the socket
pools (though H2 sessions are still layered on top of them),
there are a number of related cleanups that will need to
follow this CL.

This is part of an effort to flatten the socket pools.
https://docs.google.com/document/d/1g0EA4iDqaDhNXA_mq-YK3SlSX-xRkoKvZetAQqdRrxM/edit

Bug: 472729
Change-Id: I98aee1ac0235440902deed1e3d66479efae3af40
Reviewed-on: https://chromium-review.googlesource.com/c/1489428
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#637123}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aade58102a4ffc39c6f147ac3fa9ebe8ea4c54ef