Factor out common SameSite computations.
There are done in multiple spots, some of which are wrong, and yet more spots
will need to do them, so common functions are called for.

This also renames the enum constants to be less focused on get since set
will likely need to check them, too, and organizes them in a way to leave place
for intermediate modes (first party set relevant to ephemeral?), but this has
the cost of making read sites harder to understand, so some refinement in review
may be in order.

Change-Id: I61de22d057e13eadf396f8b324a1038b939c5f2b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1481859
Commit-Queue: Maks Orlovich <morlovich@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#637689}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 64e2913dbe703760fdef7e8ac72ec7bb729eaefa