Initialize URLRequestContextBuilder members in the header.
As is, the ifdefs in the header and the ifdefs in the initializer list
have to match. With this CL, there is no need for the initializer list
and it is clearer that members are being initialized.

Also fix lint errors.

Change-Id: Ie9437fb15578b6ae89bd4911f79914eb49a9357a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1476437
Commit-Queue: Lei Zhang <thestig@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638056}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d0ef5ca7b94a61009eec40136f3efd38c17c5e7