Remove group name from ConnectJob layer.
ConnectJobs no longer care about group names, so having them store the
name is a bit of a layering violation.

ConnectJob group names were used for a NetLog event and by the socket
pools to locate the group a ConnectJob belonged to. To resolve the first
issue, group names are now logged by a NetLog event by the socket pools
before starting each ConnectJob (which will require an update in the
net log viewer code). To resolve the second issue, Groups are now
ConnectJob::Delegates for all the ConnectJobs associated with the group.

Beyond fixing the layering violation, this CL serves two purposes: It
will make it easier to change group_name from a string to a struct, and
it will make it easier to make CommonConnectJobParameters a per-pool
object.

Bug: 927100, 533571
Change-Id: Ife570e1277ef18b4d916eb6d3ab91a9fccf2bed8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1506567
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#638632}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ce50e1d53ef866f895c86f7d59efe027bc2c1fa8