Remove proxy information from SocketPool's group_name strings.
Proxies of different but related types on the same host+port used to
be pooled together (e.g. socks4 + socks5), so this was used to prevent
group socks4 and sock5 connections together, but allow them to be in the
same pool. That behavior has since been removed, so the proxy type no
longer needs to be part of the group name.

This CL is in preparation to replace the group name with a struct.

Bug: 533571
Change-Id: I2362b369366ea1d5df3411f364f7639731b56fe6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1509636
Reviewed-by: Asanka Herath <asanka@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#639214}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4802de68c80fddacee86ca8427736b006ea1e957