Fix crash in SpdyStream::SetPriority.
When a ConnectJob is assigned to a request in the socket pool, there is
often some reassignment of ConnectJobs to stream requests. This can
result in reprioritizing ConnectJobs. If one of those ConnectJobs was
using a SpdyProxyClientSocket that is using a torn down SpdySession,
trying to change its priority would result in a null dereference.

This CL makes reprioritizing SpdyProxyClientSockets and
QuicProxyClientSockets do nothing, as the reprioritizations in this
stage are often superfluous, and doesn't really get us much in either
case.

In followup CLs, I both want to remove the unnecessary reprioritizations
and use a fixed priority of all H2/QUIC tunnels.

Bug: 940848
Change-Id: I7c646b1bfcb18977dc2837c73f7b05172d6f908d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1521279
Commit-Queue: Ryan Hamilton <rch@chromium.org>
Reviewed-by: Ryan Hamilton <rch@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#640936}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: edaf3b880a7e338f06321b2d3977e41ee55a6aa6